validator/requirements.txt
Flask
Flask-RESTful
PyJWT